Migrate to JUnit 5

+++ src/test/kotlin/days/Day12Test.kt
@@ -2,18 +2,18 @@ package days
 
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.core.Is.`is`
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 class Day12Test {
     private val day = Day12()
 
     @Test
     fun testPartOne() {
-        assertThat(day.partOne(), `is`(17 + 8))
+        assertThat(day.partOne(), `is`((17 + 8).toInt()))
     }
 
     @Test
     fun testPartTwo() {
-        assertThat(day.partTwo(), `is`(214 + 72))
+        assertThat(day.partTwo(), `is`((214 + 72).toInt()))
     }
 }
